Continue Refactoring
