+++ src/main/java/LogIn/LoggedIn/LoggedInPresenter.java
@@ -1,0 +1,11 @@
+package LogIn.LoggedIn;
+
+import DataObjects.User;
+
+import javax.swing.*;
+
+public class LoggedInPresenter extends JFrame {
+    private User user;
+
+
+}
