--- src/main/java/LogIn/LoggedInPresenter.java
@@ -1,16 +1,0 @@
-package LogIn;
-
-import Brokerage.BrokerageController;
-import Brokerage.BrokeragePresenter;
-import DataObjects.User;
-import Transaction.MakeTransactionPresenter;
-import Transaction.TransactionHistoryPresenter;
-
-import javax.swing.*;
-import java.awt.*;
-
-public class LoggedInPresenter extends JFrame {
-    private User user;
-
-
-}
